02-SRC/TD3/requirements.txt
torch>=2.0.0
numpy>=1.24.0
gymnasium>=0.29.0
wandb>=0.15.0
tqdm>=4.65.0
hockey-env>=1.0.0
imageio>=2.31.0
